Make hte CGFloat explicit.

+++ ios/StatusPanel/Extensions/UIImage.swift
@@ -101,7 +101,7 @@ extension UIImage {
 
             let value = pixel + slidingErrorWindow.removeFirst()
             slidingErrorWindow.append(0)
-            let color = value > 0.5 ? 1.0 : 0.0
+            let color = value > 0.5 ? CGFloat(1.0) : CGFloat(0.0)
             let error = (pixel - color) / 8
             for offset in mask {
                 slidingErrorWindow[offset] = slidingErrorWindow[offset] + error
